transformed responsive-menu into normal template
