+++ resources/css/pages/Privacy.scss
@@ -1,4 +1,4 @@
-.privacy page-sections {
+.privacy section:first-of-type {
   padding-top: 100px;
   padding-bottom: 50px;
 }
