+++ resources/js/index.ts
@@ -1,5 +1,5 @@
 /* global fetchInject fetch */
-import webComponentsSupported from './modules/webComponentsSupported'
+// import webComponentsSupported from './modules/webComponentsSupported'
 // get correct file name
 const revisionedFiles = fetch('/revisionedFiles').then(response => {
   return response.json()
@@ -20,25 +20,25 @@ Promise.all([revisionedFiles, fetchInjectLoaded])
   .then(results => {
     const json = results[0]
     // make sure WC are working
-    const webComponentsAvailable = new Promise((resolve) => {
-      if (!webComponentsSupported()) {
-        return fetchInject([
-          // polyfill
-          `${baseUrl}/js/webcomponents-sd-ce.js`,
-          'https://polyfill.io/v2/polyfill.min.js?features=IntersectionObserver'
-        ]).then(() => {
-          resolve()
-        })
-      }
-      resolve()
-    })
+    // const webComponentsAvailable = new Promise((resolve) => {
+    //   if (!webComponentsSupported()) {
+    //     return fetchInject([
+    //       // polyfill
+    //       `${baseUrl}/js/webcomponents-sd-ce.js`,
+    //       'https://polyfill.io/v2/polyfill.min.js?features=IntersectionObserver'
+    //     ]).then(() => {
+    //       resolve()
+    //     })
+    //   }
+    //   resolve()
+    // })
     // load critical layout components
-    const layoutComponents = fetchInject([`${baseUrl}/${json.js['js/layoutComponents.js']}`], webComponentsAvailable)
+    // const layoutComponents = fetchInject([`${baseUrl}/${json.js['js/layoutComponents.js']}`], webComponentsAvailable)
     const body = document.querySelector('body')
     // load webfont and view intro once downloaded
     fetchInject([
       'https://fonts.googleapis.com/css?family=Montserrat:700|Noto+Serif:400,400i|Source+Sans+Pro:400,600'
-    ], layoutComponents)
+    ])
       .then(() => {
         if (body !== null) {
           body.style.opacity = '1'
@@ -52,9 +52,9 @@ Promise.all([revisionedFiles, fetchInjectLoaded])
     // load responsiveMenu
     fetchInject([
       `${baseUrl}/${json.js['js/responsiveMenu.js']}`,
-      `${baseUrl}/${json.js['js/toc.js']}`,
+      // `${baseUrl}/${json.js['js/toc.js']}`,
       `${baseUrl}/${json.js['js/rest.js']}`
-    ], webComponentsAvailable)
+    ])
       .then(() => {
         document.querySelector('responsive-menu').style.display = 'block'
       })
