+++ resources/js/modules/registerServiceWorker.ts
